Permissiveness of binding
The myosin head now incorporates a permissiveness factor (0 to 1) into
the probability it will bind or unbind-from-a-weakly-bound-state. This
permissiveness factor represents the activation state of the actin
binding site and is stored within the binding site. It can be set on a
per-thin-filament basis by accessing the permissiveness property of the
thin filament.