Modified run description for easier reading by pandas
Pandas balked at having arrays of different lengths, so the z_line_func
etc have been split up to not be an array. The z_line_func is still
encoded as a string, and now the arguments to the z_line_func are
encoded as a string as well, in z_line_args.