Ignore lack of last transitions on loading from dict
Since last transitions come from actually calling transitions we don’t
have any when the sarcomere has just been instantiated. Yet we want to
log the status of the sarcomere at this point. So we add a test to
ignore the lack of last transitions in the sarcomere dictionary when
loading from such.