Track filament orientations
Now that we are randomizing the subunit that our polymers start on, in
order to introduce randomness to which myosin faces which actin and
prevent the vernier-like force increases/decreases we see during
shortening without such randomization, we need to log our filament
orientations and also to re-instantiate with those same orientations
when loading serialized sarcomeres. Man, what a wonderfully long run on
sentence.