Limit binding probability
Binding probability has been allowed to exceed one. This was not a problem when we were assuming full activation, that is this just resulted in the XB having a unity chance of binding when it was very close after diffusive movement. That is no longer the case with actin permissiveness altering the binding probability. Now we need a 0.2 actin permissiveness case to result in a maximum chance of binding of 20%. So we are just placing a limit on the probability of binding before it is multiplied by the actin permissiveness.