Powerstroke: increase max rate and shift right
As a (currently experimental) part of reforming the kinetics rates to permit positive work generation, increase the maximum rate at which bound cross bridges can undergo a power stroke from 0.2 to 0.6. The original max was arbitrarily chosen to emulate earlier models, but doesn’t have other justification. Additionally, rightward shift the offset at which the power stroke starts to kick in by ~3nm to reduce the population of loosely bound XBs that are in a drag state.