Increase and rightward shift of power stroke
Convert the power stroke from a rate/probability that works at a single timestep length to a rate and then probability that is scaled by the timestep length. This makes the power stroke robust to changes in timestep length. Additionally we increase the probability of a power stroke occurring and rightward shift it rightward with the goal of decreasing the fraction of loosely bound cross bridges that get into a drag-inducing regime without making it to the strongly bound (or detached) state.