Binding subject to same rate conversion
Recognizing that the rate to probability conversion was being done
differently for binding than for power stroke or detachment, we now
have _bind return a rate that is converted to a probability and
modified by actin permissiveness within the transition function.
Likewise p21 is now r21 and converted to p21 within the transition
function.