More on .sarc tarballs
There are still problems getting the .sarc files compressed and
uploaded from the EC2 instances. I feel like this is a silly problem
but it is actually quite hard to track down as it only seems to happen
with the results from very large runs (so no testing with toy runs) and
only on remote machines (so no local testing) and only when they are
unmonitored (so no interactive testing). This is the trifecta of
hard-to-track down problems. This change might help… maybe?