Convert lattice_spacing to prop
Lattice spacing has been converted to a property. All references to
get_lattice_spacing methods are removed. This unifies the
lattice_spacing interface and is a step towards the de-javaing of the
model. This is also a necessary step on the path to implementing a
sensible poisson ratio lattice spacing setter.