Major: poisson ratio governs ls
This is a large, non-backwards compatible, change that increments the
version number.

Lattice spacing is now changed automatically every time the z-line
location is updated according to the poisson ratio of the filament
lattice. This is declared upon creation of the lattice by the newly
passed poisson argument. This argument defaults to 0, a
sliding-filament lattice, but can be set to 0.5 (constant volume), a
negative value (auxetic), or anything in between.