Len and act out of emit_meta
emit_meta is now in its own module, metas, since it had grown too large
for run. Also, the generation of time, zline, and actin permissiveness
traces were moved to their own functions. This is to permit multiple
types of runs to be used without having to have multiple git branches.
Additionally, as part of the switch to a poisson ratio derived lattice
spacing, we no longer are using lattice spacing as an input to the meta
files.