Move tropomyosin to own file and refactor it
Tropomyosin was getting complicated to manage as a single class. I
refactored it to two classes: a tropomyosin (chain) class and a
tropomyosin site class. The site class manages the kinetics and much of
the accounting while the tropomyosin class manages spread of activation
and other larger scale phenomena. I also fixed problems with the
serialization of both of these classes (needs more testing).