Stabilize returned binding site for XB query
This is actually a potentially result shifting bug.
When a myosin head was asking for the closest binding site,
the thin face would return the correct site if it was the one
to the right of where the myosin head lay on the thin filament, but
if the correct site was to the left the thin face would return the
binding site two to the left. This gave a 'two steps forward, one
step back' form to the axial-offset vs binding-site-index graph.

I've corrected which binding site is returned and renamed the relevant
variables to hopefully make the whole thing clearer and reduce the use
of +1/-1 indexing which was the cause of this confusion in the first
place.

Note: this was discovered because it shows up as selecting an index
beyond the end of the filament for very short sarcomere lengths.