Span is calculated at the tropomyosin site level
This is to allow span of cooperative activation to vary with the tension experienced
at any given point along the thin filament. The goal is to have differing responses to
force development along the thin filament depending on how much load that location
is bearing. The nodes nearest the m-line aren't under much strain and so shouldn't see
strain dependent effects, while those nearest the z-disk should see a lot.